fix: wrong params name of function encode in ios new arch

+++ src/specs/NativeBlurhashModule.ts
@@ -1,8 +1,8 @@
 import { type TurboModule, TurboModuleRegistry } from 'react-native';
-import type { Int32 } from 'react-native/Libraries/Types/CodegenTypes';
+import type { Double } from 'react-native/Libraries/Types/CodegenTypes';
 
 export interface Spec extends TurboModule {
-	createBlurhashFromImage: (imageUri: string, componentsX: Int32, componentsY: Int32) => Promise<string>;
+	createBlurhashFromImage: (imageUri: string, componentsX: Double, componentsY: Double) => Promise<string>;
 	clearCosineCache: () => void;
 }
